move saveGenerator to utils
