+++ src/views/Generator/Generator.tsx
@@ -2,7 +2,7 @@ import { Allotment } from 'allotment'
 import { Button } from '@radix-ui/themes'
 import { useEffect } from 'react'
 
-import { exportScript, saveScript } from './Generator.utils'
+import { exportScript, saveScript, saveGenerator } from './Generator.utils'
 import { PageHeading } from '@/components/Layout/PageHeading'
 import { harToProxyData } from '@/utils/harToProxyData'
 import { GeneratorDrawer } from './GeneratorDrawer'
@@ -10,11 +10,6 @@ import { GeneratorSidebar } from './GeneratorSidebar'
 import { useGeneratorStore } from '@/hooks/useGeneratorStore'
 import { TestRuleContainer } from './TestRuleContainer'
 import { AllowList } from './AllowList/AllowList'
-import {
-  GeneratorFile,
-  GeneratorOptions,
-  GeneratorTestData,
-} from '@/types/generator'
 
 export function Generator() {
   const {
@@ -46,44 +41,6 @@ export function Generator() {
     const script = await exportScript(filteredRequests, rules)
 
     saveScript(script)
-  }
-
-  const saveGenerator = async () => {
-    const generatorState = useGeneratorStore.getState()
-    const options: GeneratorOptions = {
-      loadProfile: {
-        executor: generatorState.executor,
-        startTime: generatorState.startTime,
-        gracefulStop: generatorState.gracefulStop,
-        stages: generatorState.stages,
-        gracefulRampDown: generatorState.gracefulRampDown,
-        startVUs: generatorState.startVUs,
-        iterations: generatorState.iterations,
-        maxDuration: generatorState.maxDuration,
-        vus: generatorState.vus,
-      },
-      thinkTime: {
-        sleepType: generatorState.sleepType,
-        timing: generatorState.timing,
-      },
-    }
-    const generatorTestData: GeneratorTestData = {
-      variables: generatorState.variables,
-    }
-
-    const generatorFile: GeneratorFile = {
-      name: generatorState.name,
-      version: '0',
-      recordingPath: generatorState.recordingPath,
-      options: options,
-      testData: generatorTestData,
-      rules: generatorState.rules,
-      allowlist: generatorState.allowList,
-    }
-
-    window.studio.generator.saveGenerator(
-      JSON.stringify(generatorFile, null, 2)
-    )
   }
 
   return (
