+++ src/views/Generator/Generator.utils.ts
@@ -6,10 +6,12 @@ import * as prettierPluginBabel from 'prettier/plugins/babel'
 // eslint-disable-next-line import/namespace
 import * as prettierPluginEStree from 'prettier/plugins/estree'
 import { groupProxyData } from '@/utils/groups'
-import { GeneratorState } from '@/hooks/useGeneratorStore/types'
-import { CommonOptions } from './GeneratorDrawer/LoadProfile/types'
-import { ExecutorType } from '@/constants/generator'
-import { exhaustive } from '@/utils/typescript'
+import { useGeneratorStore } from '@/hooks/useGeneratorStore'
+import {
+  GeneratorOptions,
+  GeneratorTestData,
+  GeneratorFile,
+} from '@/types/generator'
 
 export async function exportScript(recording: ProxyData[], rules: TestRule[]) {
   const groupedProxyData = groupProxyData(recording)
@@ -29,31 +31,38 @@ export function saveScript(script: string) {
   window.studio.script.saveScript(script)
 }
 
-export const getLoadProfile = (state: GeneratorState) => {
-  const commonOptions: CommonOptions = {
-    executor: state.executor,
-    startTime: state.startTime,
-    gracefulStop: state.gracefulStop,
+export const saveGenerator = () => {
+  const generatorState = useGeneratorStore.getState()
+  const options: GeneratorOptions = {
+    loadProfile: {
+      executor: generatorState.executor,
+      startTime: generatorState.startTime,
+      gracefulStop: generatorState.gracefulStop,
+      stages: generatorState.stages,
+      gracefulRampDown: generatorState.gracefulRampDown,
+      startVUs: generatorState.startVUs,
+      iterations: generatorState.iterations,
+      maxDuration: generatorState.maxDuration,
+      vus: generatorState.vus,
+    },
+    thinkTime: {
+      sleepType: generatorState.sleepType,
+      timing: generatorState.timing,
+    },
+  }
+  const generatorTestData: GeneratorTestData = {
+    variables: generatorState.variables,
   }
 
-  switch (state.executor) {
-    case ExecutorType.RampingVUs:
-      return {
-        stages: state.stages,
-        gracefulRampDown: state.gracefulRampDown,
-        startVUs: state.startVUs,
-        ...commonOptions,
-        executor: ExecutorType.RampingVUs,
-      }
-    case ExecutorType.SharedIterations:
-      return {
-        iterations: state.iterations,
-        maxDuration: state.maxDuration,
-        vus: state.vus,
-        ...commonOptions,
-        executor: ExecutorType.SharedIterations,
-      }
-    default:
-      return exhaustive(state.executor)
+  const generatorFile: GeneratorFile = {
+    name: generatorState.name,
+    version: '0',
+    recordingPath: generatorState.recordingPath,
+    options: options,
+    testData: generatorTestData,
+    rules: generatorState.rules,
+    allowlist: generatorState.allowList,
   }
+
+  window.studio.generator.saveGenerator(JSON.stringify(generatorFile, null, 2))
 }
